requirements.txt
# UI
streamlit

# PDF parsing
PyPDF2

# Embedding & LLM
cohere
python-dotenv

# Vector store
faiss-cpu

# Optional: agar tidak error di Windows
numpy
